+++ NEWS.md
@@ -10,6 +10,8 @@
 
 - `default_kind()` now looks for the archetype matching the content. For `/content/post/2018-02-24-postslug.Rmd` it looks for `archetypes/post.md` (thanks, @lcolladotor, #173).
 
+- `new_content()` now works with files that end in .Rmd and .Rmarkdown. The archetype still has to end in .md for hugo to work with it (thanks, @lcolladotor, #261).
+
 # CHANGES IN blogdown VERSION 0.5
 
 ## BUG FIXES