+++ NEWS.md
@@ -18,6 +18,10 @@
 
 - The `kind` argument (i.e., the archetype) of `new_content()` now works with files that end in `.Rmd` and `.Rmarkdown`. The archetype still has to end in `.md` for Hugo to work with it (thanks, @lcolladotor, #261).
 
+## MINOR CHANGES
+
+- The Github repo `yihui/hugo-lithium-theme` was renamed to `yihui/hugo-lithiutm`, and the default `theme` argument value was changed accordingly (thanks, @rorynolan, #291).
+
 # CHANGES IN blogdown VERSION 0.5
 
 ## BUG FIXES